pyproject.toml
[project]
name = "docusense"
version = "0.1.0"
authors = [{ name = "Andre garasic", email = "andre.garasic@gmail.com" }]
description = "DocuSense - A FastAPI REST API application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.4",
    "alembic>=1.13.0",
    # File uploads
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    # Embeddings
    "sentence-transformers>=2.2.0",
    # Text extraction
    "pymupdf>=1.24.0",
    "easyocr>=1.7.0",
    "pillow>=10.0.0",
    # QA model
    "transformers>=4.36.0",
    "torch>=2.1.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Caching
    "cachetools>=5.3.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["docusense"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
